BVH refactoring (#7)
* add empty BVH instance (often used as initial value when no entities/primitives are loaded yet)

* refactor BVH instantiation into easier builder pattern
- removes BVHImplementation and moves interface to BVH
- replace constructor calls by factory calls

* move BVH to own package and introduce Intersectable interface (suggesting use for e.g. octree later)
