+++ chunky/src/java/se/llbit/chunky/ui/render/Poser.java
@@ -36,19 +36,17 @@ import org.apache.commons.math3.util.FastMath;
 import se.llbit.chunky.renderer.scene.Camera;
 import se.llbit.chunky.renderer.scene.PlayerModel;
 import se.llbit.chunky.ui.DoubleAdjuster;
-import se.llbit.math.BVH;
+import se.llbit.math.bvh.BVH;
 import se.llbit.math.ColorUtil;
 import se.llbit.math.Matrix3;
 import se.llbit.math.QuickMath;
 import se.llbit.math.Ray;
 import se.llbit.math.Vector3;
-import se.llbit.util.TaskTracker;
 
 import java.io.File;
 import java.io.IOException;
 import java.net.URL;
 import java.nio.IntBuffer;
-import java.util.Collections;
 import java.util.ResourceBundle;
 
 /**
@@ -58,7 +56,7 @@ public class Poser extends Stage implements Initializable {
   private static final WritablePixelFormat<IntBuffer> PIXEL_FORMAT =
       PixelFormat.getIntArgbInstance();
   private final EntitiesTab.EntityData player;
-  private BVH bvh = new BVH(Collections.emptyList(), new Vector3(), TaskTracker.Task.NONE);
+  private BVH bvh = BVH.EMPTY;
   private int[] pixels;
   private int width = 300;
   private int height = 300;
