+++ chunky/src/java/se/llbit/math/bvh/SahBVH.java
@@ -15,11 +15,12 @@
  * You should have received a copy of the GNU General Public License
  * along with Chunky.  If not, see <http://www.gnu.org/licenses/>.
  */
-package se.llbit.math;
+package se.llbit.math.bvh;
 
 import se.llbit.chunky.entity.Entity;
 import se.llbit.chunky.main.Chunky;
 import se.llbit.log.Log;
+import se.llbit.math.Vector3;
 import se.llbit.math.primitive.MutableAABB;
 import se.llbit.math.primitive.Primitive;
 import se.llbit.util.TaskTracker;
@@ -27,13 +28,11 @@ import se.llbit.util.TaskTracker;
 import java.util.*;
 import java.util.function.IntConsumer;
 
-import static se.llbit.math.BVH.SPLIT_LIMIT;
-
-public class SahMaBVH extends BinaryBVH {
+public class SahBVH extends BinaryBVH {
     public static void initImplementation() {
-        BVH.factories.put("SAH_MA", new BVH.ImplementationFactory() {
+        Factory.implementations.put("SAH", new Factory.BVHBuilder() {
             @Override
-            public BVH.BVHImplementation create(Collection<Entity> entities, Vector3 worldOffset, TaskTracker.Task task) {
+            public BVH create(Collection<Entity> entities, Vector3 worldOffset, TaskTracker.Task task) {
                 task.update(1000, 0);
                 double entityScaler = 500.0 / entities.size();
                 int done = 0;
@@ -49,20 +48,20 @@ public class SahMaBVH extends BinaryBVH {
                 primitives = null; // Allow the collection to be garbage collected during construction when only the array is used
 
                 double primitiveScaler = 500.0 / allPrimitives.length;
-                return new SahMaBVH(allPrimitives, i -> task.updateInterval((int) (i * primitiveScaler) + 500, 1));
+                return new SahBVH(allPrimitives, i -> task.updateInterval((int) (i * primitiveScaler) + 500, 1));
             }
 
             @Override
             public String getTooltip() {
-                return "Fast and nearly optimal BVH building method.";
+                return "Slow but nearly optimal BVH building method.";
             }
         });
     }
 
-    public SahMaBVH(Primitive[] primitives, IntConsumer task) {
-        Node root = constructSAH_MA(primitives, task);
+    public SahBVH(Primitive[] primitives, IntConsumer task) {
+        Node root = constructSAH(primitives, task);
         pack(root);
-        Log.info("Built SAH_MA BVH with depth " + this.depth);
+        Log.info("Built SAH BVH with depth " + this.depth);
     }
 
     private enum Action {
@@ -71,10 +70,9 @@ public class SahMaBVH extends BinaryBVH {
     }
 
     /**
-     * Construct a BVH using Surface Area Heuristic (SAH)
-     * This splits along the major axis which usually gets good results.
+     * Construct a BVH using Surface Area Heuristic (SAH).
      */
-    private Node constructSAH_MA(Primitive[] primitives, IntConsumer task) {
+    private Node constructSAH(Primitive[] primitives, IntConsumer task) {
         int progress = 0;
 
         Stack<Node> nodes = new Stack<>();
@@ -94,7 +92,7 @@ public class SahMaBVH extends BinaryBVH {
                     progress += chunk.length;
                     task.accept(progress);
                 } else {
-                    splitSAH_MA(chunk, actions, chunks);
+                    splitSAH(chunk, actions, chunks);
                 }
             }
         }
@@ -102,22 +100,9 @@ public class SahMaBVH extends BinaryBVH {
     }
 
     /**
-     * Split a chunk based on Surface Area Heuristic of all possible splits on the major axis.
+     * Split a chunk based on Surface Area Heuristic of all possible splits
      */
-    private void splitSAH_MA(Primitive[] chunk, Stack<Action> actions, Stack<Primitive[]> chunks) {
-        AABB bb = bb(chunk);
-        double xl = bb.xmax - bb.xmin;
-        double yl = bb.ymax - bb.ymin;
-        double zl = bb.zmax - bb.zmin;
-        Comparator<Primitive> cmp;
-        if (xl >= yl && xl >= zl) {
-            cmp = cmpX;
-        } else if (yl >= xl && yl >= zl) {
-            cmp = cmpY;
-        } else {
-            cmp = cmpZ;
-        }
-
+    private void splitSAH(Primitive[] chunk, Stack<Action> actions, Stack<Primitive[]> chunks) {
         MutableAABB bounds = new MutableAABB(0, 0, 0, 0, 0, 0);
         double cmin = Double.POSITIVE_INFINITY;
         int split = 0;
@@ -126,7 +111,8 @@ public class SahMaBVH extends BinaryBVH {
         double[] sl = new double[end];
         double[] sr = new double[end];
 
-        Chunky.getCommonThreads().submit(() -> Arrays.parallelSort(chunk, cmp)).join();
+        Comparator<Primitive> cmp = cmpX;
+        Chunky.getCommonThreads().submit(() -> Arrays.parallelSort(chunk, cmpX)).join();
         for (int i = 0; i < end - 1; ++i) {
             bounds.expand(chunk[i].bounds());
             sl[i] = bounds.surfaceArea();
@@ -142,6 +128,49 @@ public class SahMaBVH extends BinaryBVH {
                 cmin = c;
                 split = i;
             }
+        }
+
+        Chunky.getCommonThreads().submit(() -> Arrays.parallelSort(chunk, cmpY)).join();
+        for (int i = 0; i < end - 1; ++i) {
+            bounds.expand(chunk[i].bounds());
+            sl[i] = bounds.surfaceArea();
+        }
+        bounds = new MutableAABB(0, 0, 0, 0, 0, 0);
+        for (int i = end - 1; i > 0; --i) {
+            bounds.expand(chunk[i].bounds());
+            sr[i - 1] = bounds.surfaceArea();
+        }
+        for (int i = 0; i < end - 1; ++i) {
+            double c = sl[i] * (i + 1) + sr[i] * (end - i - 1);
+            if (c < cmin) {
+                cmin = c;
+                split = i;
+                cmp = cmpY;
+            }
+        }
+
+        Chunky.getCommonThreads().submit(() -> Arrays.parallelSort(chunk, cmpZ)).join();
+        for (int i = 0; i < end - 1; ++i) {
+            bounds.expand(chunk[i].bounds());
+            sl[i] = bounds.surfaceArea();
+        }
+        bounds = new MutableAABB(0, 0, 0, 0, 0, 0);
+        for (int i = end - 1; i > 0; --i) {
+            bounds.expand(chunk[i].bounds());
+            sr[i - 1] = bounds.surfaceArea();
+        }
+        for (int i = 0; i < end - 1; ++i) {
+            double c = sl[i] * (i + 1) + sr[i] * (end - i - 1);
+            if (c < cmin) {
+                cmin = c;
+                split = i;
+                cmp = cmpZ;
+            }
+        }
+
+        if (cmp != cmpZ) {
+            Comparator<Primitive> finalCmp = cmp;
+            Chunky.getCommonThreads().submit(() -> Arrays.parallelSort(chunk, finalCmp)).join();
         }
 
         split += 1;
